changing junit to use general aggregate logic
