+++ plugin/util.go
@@ -7,9 +7,14 @@ package plugin
 import (
 	"context"
 	"encoding/json"
+	"encoding/xml"
 	"fmt"
+	"github.com/bmatcuk/doublestar/v4"
 	influxdb2 "github.com/influxdata/influxdb-client-go/v2"
+	"github.com/sirupsen/logrus"
+	"io"
 	"os"
+	"path/filepath"
 	"strings"
 	"time"
 )
@@ -39,11 +44,78 @@ type DbCredentials struct {
 	Bucket        string
 }
 
+func GetXmlReportData[T any](reportsRootDir string, patterns []string) ([]T, error) {
+
+	fmt.Println("GetXmlReportData: reportsRootDir ==  ", reportsRootDir)
+
+	var xmlReportFiles []string
+	var xmlFileReportDataList []T
+
+	for _, pattern := range patterns {
+		fmt.Println("junit: pattern ==  ", pattern)
+		fmt.Println("junit: reportsRootDir ==  ", reportsRootDir)
+		tmpReportDir := os.DirFS(reportsRootDir)
+		relPattern := strings.TrimPrefix(pattern, reportsRootDir+"/")
+		filesList, err := doublestar.Glob(tmpReportDir, relPattern)
+		if err != nil {
+			logrus.Println("Include patterns not found ", err.Error())
+			return xmlFileReportDataList, err
+		}
+		xmlReportFiles = append(xmlReportFiles, filesList...)
+	}
+
+	fmt.Println("xmlReportFiles ==  ", xmlReportFiles)
+	fmt.Println("len(xmlReportFiles) ", len(xmlReportFiles))
+
+	for _, xmlReportFile := range xmlReportFiles {
+		fmt.Println("Processing junit result file: ", xmlReportFile)
+		tmpXmlReportFile := filepath.Join(reportsRootDir, xmlReportFile)
+		report := ParseXmlReport[T](tmpXmlReportFile)
+		reportBytes, err := json.Marshal(report)
+		if err != nil {
+			fmt.Println("Error marshalling report: %v", err)
+			logrus.Println("Error marshalling report: %v", err)
+		}
+
+		xmlFileReport, err := ToStructFromJsonString[T](string(reportBytes))
+		if err != nil {
+			fmt.Println("Error converting json to struct: %v", err)
+			logrus.Println("Error converting json to struct: %v", err)
+			return xmlFileReportDataList, err
+		}
+
+		xmlFileReportDataList = append(xmlFileReportDataList, xmlFileReport)
+	}
+
+	return xmlFileReportDataList, nil
+}
+
+func ParseXmlReport[T any](filename string) T {
+	file, err := os.Open(filename)
+	if err != nil {
+		fmt.Println("Error opening XML file: ", err)
+		logrus.Fatalf("Error opening XML file: %v", err)
+	}
+	defer file.Close()
+
+	data, err := io.ReadAll(file)
+	if err != nil {
+		fmt.Println("Error reading XML file ", err)
+	}
+
+	var report T
+	err = xml.Unmarshal(data, &report)
+	if err != nil {
+		fmt.Println("Error unmarshalling XML: %v", err)
+	}
+	return report
+}
+
 func Aggregate[T any](reportsDir, includes string,
 	dbUrl, dbToken, dbOrg, dbBucket, measurementName string,
 	calculateAggregate func(testNgAggregatorList []T) T,
 	getDataMaps func(pipelineId, buildNumber string,
-	aggregateData T) (map[string]string, map[string]interface{})) error {
+		aggregateData T) (map[string]string, map[string]interface{})) error {
 
 	reportsRootDir := reportsDir
 	patterns := strings.Split(includes, ",")
